Move Binding class out of pipeline.py
